requirements.txt
autopep8==1.6.0
click>=8.0.4
Flask>=2.0.3
importlib-metadata>=4.8.3
itsdangerous>=2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
uWSGI==2.0.17.1
pycodestyle==2.8.0
PyYAML==6.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.0